% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_predicted_ancestry}
\alias{plot_predicted_ancestry}
\title{Plot projected PCs by predicted ancestral superpopulation, and compare with 1kg reference}
\usage{
plot_predicted_ancestry(
  projected_pcs = new_pcs,
  projected_ancestry = "ancestry_estimates.tsv",
  outfile = "ancestry_estimates_plot"
)
}
\arguments{
\item{projected_pcs}{a data frame with transformed PC scores. Default output from transform_pcs().}

\item{projected_ancestry}{a path to a .tsv file with ancestry estimates per person, as generated by predict_ancestry().}

\item{outfile}{a path for the output plot}
}
\value{
a plot which will display the projected PC values for each individual in your sample, grouped by predicted ancestral superpopulation of origin, alongside reference data from 1kg in the same PC co-ordinate space
}
\description{
Plot projected PCs by predicted ancestral superpopulation, and compare with 1kg reference
}
\examples{
#' # First - project PCs:
project_pcs(out = "projected_pcs")
# Then run:
new_pcs = transform_pcs("projected_pcs.sscore")
# Then plot with:
plot_with_1kg(projected_pcs = new_pcs, save_plot = TRUE, outfile = "niceplot")
# Predict ancestry with:
predict_ancestry(new_pcs, outfile = "ancestry_estimates")
# Now plot predictions alongside 1kg superpopulations for reference
plot_predicted_ancestry(projected_pcs = new_pcs, projected_ancestry = "ancestry_estimates.tsv")
}
