% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{project_pcs}
\alias{project_pcs}
\title{Project PCs}
\usage{
project_pcs(
  bfile = "ukb_merged_chroms",
  out = "projected_pcs",
  n_threads = 10,
  path_to_plink_exe = "module unload plink; module load plink/2.0-20200328; plink2 "
)
}
\arguments{
\item{bfile}{a plink1 binary fileset containing genotypes for individuals whose ancestral superpopulation you want to estimate.}

\item{out}{a path to the output file}

\item{path_to_plink_exe}{the path to your plink2 executable}
}
\value{
a '.sscore' file with the projected PC values for individuals in your dataset in 1kg PC space.
}
\description{
Project PCs
}
