% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_pcs}
\alias{transform_pcs}
\title{Transform projected PCs onto 1kg scale}
\usage{
transform_pcs(projected_pcs = "projected_pcs.sscore")
}
\arguments{
\item{projected_pcs}{}
}
\value{
transformed PC scores on the same co-ordinates as the 1kg reference samples
}
\description{
Transform projected PCs onto 1kg scale
}
