% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_predicted_ancestry}
\alias{plot_predicted_ancestry}
\title{Title}
\usage{
plot_predicted_ancestry(
  projected_pcs = newpcs,
  projected_ancestry = pred_output
)
}
\arguments{
\item{projected_ancestry}{}
}
\value{

}
\description{
Title
}
