% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_with_1kg}
\alias{plot_with_1kg}
\title{Title Plot PCs with 1kg reference}
\usage{
plot_with_1kg(projected_pcs = newpcs)
}
\arguments{
\item{projected_pcs}{projected PC scores - a '.sscore' file produced by project_pcs & transformed with transform_pcs}
}
\value{
a plot with the new samples projected on 1kg PC space
}
\description{
Title Plot PCs with 1kg reference
}
