% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_pcs}
\alias{transform_pcs}
\title{Transform projected PCs onto 1kg scale}
\usage{
transform_pcs(projected_pcs = "projected_pcs.sscore")
}
\arguments{
\item{projected_pcs}{transformed PC scores on the same co-ordinates as the 1kg reference samples. Default output from project_pcs().}
}
\value{
A data frame with raw PC scores on the same co-ordinates as the 1kg reference samples. Default output from project_pcs().
}
\description{
Transform projected PCs onto 1kg scale
}
\examples{

# First - project PCs:
project_pcs(out = "projected_pcs")
# Then run:
new_pcs = transform_pcs("projected_pcs.sscore")
}
